requirements.txt
boto3==1.34.29
requests==2.31.0
ratelimit==2.2.1
smart-open==7.0.1
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
tqdm==4.66.1
psutil>=5.9.0
matplotlib==3.8.2
aws-glue-sessions==1.0.0
pyspark==3.4.1
pandas>=1.5.0
numpy>=1.23.0
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0
aiohttp>=3.9.0
aiofiles>=23.2.1
aiodns>=3.1.1
cachetools>=5.3.0
aioredis>=2.0.0
redis>=5.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-aiohttp-client>=0.41b0
prometheus-client>=0.17.0
opentelemetry-exporter-otlp>=1.20.0
scikit-learn>=1.3.0
pyjwt>=2.8.0
cryptography>=41.0.0
pandas>=2.0.0 